+++ test/audit.messages.ts
@@ -5,7 +5,7 @@ const custom_fz = (custom: any) => custom.z = 2
 
 export default {
   print: false,
-  pattern: 'sys:telemetry',
+  pattern: 'sys:audit',
   allow: { missing: true },
 
   calls: [
